RF-22: addressing changes requested
