+++ src/types/navigation.ts
@@ -1,0 +1,17 @@
+import { NavigatorScreenParams } from '@react-navigation/native';
+
+export type BottomTabParamList = {
+  RoutinesScreen: undefined;
+  HomeScreen: undefined;
+  ExerciseScreen: undefined;
+};
+
+export type MainDrawerParamList = {
+  Home: undefined;
+  Profile: undefined;
+};
+
+export type RootStackParamList = {
+  // TODO : onboarding , auth, etc
+  Main: NavigatorScreenParams<MainDrawerParamList>;
+};
